@@ -9,35 +9,44 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@react-google-maps/api": "^2.20.3",
+    "@googlemaps/js-api-loader": "^1.16.8",
+    "@react-google-maps/api": "^2.20.5",
+    "@types/imap-simple": "^4.2.9",
     "@types/mongoose": "^5.11.97",
+    "@types/nodemailer": "^6.4.17",
     "bcryptjs": "^2.4.3",
     "browser-image-compression": "^2.0.2",
+    "imap-simple": "^1.6.3",
     "jwt-decode": "^4.0.0",
     "moment-timezone": "^0.5.46",
     "mongoose": "^8.8.3",
     "next": "^15.1.0",
+    "nodemailer": "^6.9.16",
     "pino": "^9.5.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-google-recaptcha": "^3.1.0",
     "react-hot-toast": "^2.4.1",
+    "react-modal": "^3.16.3",
     "react-text-mask": "^5.5.0",
+    "text-mask-addons": "^3.8.0",
     "uuid": "^11.0.3"
   },
   "devDependencies": {
     "@types/bcryptjs": "^2.4.6",
+    "@types/google.maps": "^3.58.1",
     "@types/jsonwebtoken": "^9.0.7",
-    "@types/node": "^20",
+    "@types/node": "^20.17.12",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "@types/react-google-recaptcha": "^2.1.9",
+    "@types/react-modal": "^3.16.3",
     "@types/react-text-mask": "^5.4.14",
     "eslint": "^8",
     "eslint-config-next": "15.0.3",
     "jsonwebtoken": "^9.0.2",
     "postcss": "^8",
     "tailwindcss": "^3.4.16",
-    "typescript": "^5"
+    "typescript": "^5.7.3"
   }
 }
