@@ -18,7 +18,7 @@ const geistMono = localFont({
 });
 
 export const metadata: Metadata = {
-  title: "PG Mechanical Demo",
+  title: "PG Mechanical App",
   description: "Generated by create next app",
 };
 
