@@ -1,5 +1,5 @@
-export interface ITicket {
-  _id?: string;
+export interface Ticket {
+  _id: string;
   ticketNumber: string;
   name: string;
   email: string;
@@ -8,6 +8,7 @@ export interface ITicket {
   workOrderDescription: string;
   timeAvailability: string;
   status: string;
+  inProgress: boolean;
   assignedTo?: string;
   createdAt: Date;
   updatedAt: Date;
@@ -17,4 +18,50 @@ export interface ITicket {
   amountBilled?: number;
   amountPaid?: string;
   images?: { type: string, default: [] }[];
+  priority?: Priority;
 }
+
+export interface Parts {
+  category: string;
+  parts: string[];
+}
+
+export interface User {
+  username: string;
+  password: string;
+  is_admin: boolean;
+}
+
+export interface TimeRange {
+  startTime: Date;
+  endTime?: Date;
+}
+
+export interface TimeEntry {
+  user: string;
+  ticket: string;
+  timeRanges: TimeRange[];
+  week: number;
+}
+
+export interface UserHours {
+  [user: string]: {
+    total: number;
+    weekly: {
+      [week: number]: number;
+    };
+  };
+}
+
+export type Priority = 'Highest' | 'High' | 'Medium' | 'Low' | 'Lowest' | '';
+
+export type ModalType = 'none' | 'priority' | 'notes' | 'popup' | 'confirmation';
+
+export const priorityMap: { [key in Priority]: number | null } = {
+  'Highest': 1,
+  'High': 2,
+  'Medium': 3,
+  'Low': 4,
+  'Lowest': 5,
+  '': null,
+};
