@@ -1,6 +1,7 @@
 "use client";
 import React, { useState, useEffect, useRef } from 'react';
 import { useRouter } from 'next/navigation';
+import Image from 'next/image';
 import { useAuth } from '@/context/AuthContext';
 
 export default function Header() {
@@ -31,7 +32,9 @@ export default function Header() {
 
   return (
     <div className="bg-gradient-cycling flex justify-between items-center p-4 bg-gray-800 text-white">
-      <h1 className="text-xl font-bold text-black">PG Mechanical Board</h1>
+      <header>
+        <Image src="/logo.png" alt="Logo" width={64} height={64} className='w-16 h-16' onClick={() => router.push('/')} />      </header>
+      <h1 className="text-3xl font-bold text-black">PG Mechanical</h1>
       <div className="relative" ref={dropdownRef}>
         {isLoggedIn ? (
           <button
@@ -51,23 +54,54 @@ export default function Header() {
           </button>
         )}
         {dropdownOpen && isLoggedIn && (
-          <div className="absolute right-0 mt-2 w-48 bg-white border border-gray-300 rounded shadow-lg">
+          <div className="absolute right-0 mt-2 w-48 bg-white border border-gray-300 rounded shadow-lg z-50">
             {isAdmin && (
               <button
-                onClick={() => router.push('/create-user')}
+                onClick={() => {
+                  setDropdownOpen(false);
+                  router.push('/create-user');
+                }}
                 className="block w-full text-left px-4 py-2 text-gray-700 hover:bg-gray-100"
               >
                 Create User
               </button>
             )}
+            {isAdmin && (
+              <button
+                onClick={() => {
+                  setDropdownOpen(false);
+                  router.push('/time-logs');
+                }}
+                className="block w-full text-left px-4 py-2 text-gray-700 hover:bg-gray-100"
+              >
+                Time Logs
+              </button>
+            )}
+            {isAdmin && (
+              <button
+                onClick={() => {
+                  setDropdownOpen(false);
+                  router.push('/all-users');
+                }}
+                className="block w-full text-left px-4 py-2 text-gray-700 hover:bg-gray-100"
+              >
+                All Users
+              </button>
+            )}
             <button
-              onClick={() => router.push('/tickets')}
+              onClick={() => {
+                setDropdownOpen(false);
+                router.push('/tickets');
+              }}
               className="block w-full text-left px-4 py-2 text-gray-700 hover:bg-gray-100"
             >
               Tickets
             </button>
             <button
-              onClick={handleSignOut}
+              onClick={() => {
+                setDropdownOpen(false);
+                handleSignOut();
+              }}
               className="block w-full text-left px-4 py-2 text-gray-700 hover:bg-gray-100"
             >
               Sign Out
