@@ -1,10 +1,10 @@
 import { Schema, model, Document } from 'mongoose';
-import { ITicket } from '@/common/interfaces';
+import { Ticket } from '@/common/interfaces';
 import { TICKET_STATUSES } from '@/common/constants';
 
-interface ITicketDocument extends Omit<ITicket, '_id'>, Document {}
+interface ITicket extends Omit<Ticket, '_id'>, Document { }
 
-const ticketSchema = new Schema<ITicketDocument>({
+const ticketSchema = new Schema<ITicket>({
   ticketNumber: { type: String, required: true, unique: true },
   name: { type: String, required: true },
   email: { type: String, required: true },
@@ -13,17 +13,19 @@ const ticketSchema = new Schema<ITicketDocument>({
   workOrderDescription: { type: String, required: true },
   timeAvailability: { type: String, required: true },
   status: { type: String, required: true, enum: TICKET_STATUSES, default: 'New' },
+  inProgress: { type: Boolean, required: true, default: false },
   assignedTo: { type: String },
   createdAt: { type: Date, default: Date.now },
   updatedAt: { type: Date, default: Date.now },
-  partsUsed: {type : [String], default: []},
-  servicesDelivered: {type : String, default: ''},
-  additionalNotes: {type : String, default: ''},
-  amountBilled : {type : Number, default: 0},
-  amountPaid: {type : Number, default: 0},
+  partsUsed: { type: [String], default: [] },
+  servicesDelivered: { type: String, default: '' },
+  additionalNotes: { type: String, default: '' },
+  amountBilled: { type: Number, default: 0 },
+  amountPaid: { type: Number, default: 0 },
   images: [{ type: String, default: [] }],
-});
+  priority: { type: String, enum: ['', 'Highest', 'High', 'Medium', 'Low', 'Lowest'], default: '' },
+}, { collection: 'tickets' });
 
-const Ticket = model<ITicketDocument>('Ticket', ticketSchema);
+const TicketModel = model<ITicket>('Ticket', ticketSchema);
 
-export default Ticket;
+export default TicketModel;
