@@ -1,17 +1,14 @@
 import { Schema, model, Document } from 'mongoose';
+import { User } from '@/common/interfaces';
 
-interface IUser extends Document {
-  username: string;
-  password: string;
-  phone_number: string;
-}
+interface IUser extends User, Document { }
 
 const userSchema = new Schema<IUser>({
   username: { type: String, required: true, unique: true },
   password: { type: String, required: true },
-  phone_number: { type: String, required: true },
-});
+  is_admin: { type: Boolean, required: true },
+}, { collection: 'users' });
 
-const User = model<IUser>('User', userSchema);
+const UserModel = model<IUser>('User', userSchema);
 
-export default User;
+export default UserModel;
